+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
  - CSS classes to DataTableHelper.
  - Methods to StringHelper class.
+ - Methods to ListHelper class.
 
 ## [1.4.0] - 2019-01-12
 ### Added